refactor: decompose RightPanel into smaller modules
Split the monolithic RightPanel.tsx (366 lines) into focused components
and hooks: ResizeHandle, TabsHeader, useDocumentViewer, useExpandableCards,
and shared constants. Removed unused RightPanelContext.tsx.
