+++ lexwebapp/src/components/right-panel/ResizeHandle.tsx
@@ -1,0 +1,46 @@
+import { useCallback, useEffect, useRef } from 'react';
+import { useUIStore } from '../../stores';
+
+export function ResizeHandle() {
+  const rightPanelWidth = useUIStore(s => s.rightPanelWidth);
+  const setRightPanelWidth = useUIStore(s => s.setRightPanelWidth);
+  const isResizing = useRef(false);
+  const startX = useRef(0);
+  const startWidth = useRef(0);
+
+  const handleResizeStart = useCallback((e: React.MouseEvent) => {
+    e.preventDefault();
+    isResizing.current = true;
+    startX.current = e.clientX;
+    startWidth.current = rightPanelWidth;
+    document.body.style.cursor = 'col-resize';
+    document.body.style.userSelect = 'none';
+  }, [rightPanelWidth]);
+
+  useEffect(() => {
+    const handleMouseMove = (e: MouseEvent) => {
+      if (!isResizing.current) return;
+      const delta = startX.current - e.clientX;
+      setRightPanelWidth(startWidth.current + delta);
+    };
+    const handleMouseUp = () => {
+      if (!isResizing.current) return;
+      isResizing.current = false;
+      document.body.style.cursor = '';
+      document.body.style.userSelect = '';
+    };
+    document.addEventListener('mousemove', handleMouseMove);
+    document.addEventListener('mouseup', handleMouseUp);
+    return () => {
+      document.removeEventListener('mousemove', handleMouseMove);
+      document.removeEventListener('mouseup', handleMouseUp);
+    };
+  }, [setRightPanelWidth]);
+
+  return (
+    <div
+      onMouseDown={handleResizeStart}
+      className="absolute left-0 top-0 bottom-0 w-1 cursor-col-resize hover:bg-claude-text/20 active:bg-claude-text/30 transition-colors z-50 hidden lg:block"
+    />
+  );
+}
