+++ lexwebapp/src/components/right-panel/index.ts
@@ -1,0 +1,1 @@
+export { RightPanel } from './RightPanel';
